# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OrgFlow API
  version: 1.0.0
  description: Employee and leave management. All routes except login and password reset need a Bearer JWT.
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
...
security:
  - bearer: []
paths:
...
  /auth/login:
    post:
      summary: Sign in
      security: []
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email, password], properties: { email: { type: string }, password: { type: string } } } } } }
      responses: { '200': { description: "{ token, user }" }, '401': { description: Incorrect email or password }, '429': { description: Too many attempts } }
...
  /auth/reset-password:
    post: { summary: Set a new password with an emailed token (single use, 1 hour), security: [], responses: { '200': { description: updated }, '400': { description: invalid or expired token } } }
...
  /departments/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    put: { summary: Update (admin), responses: { '200': { description: updated } } }
    delete: { summary: Delete when empty (admin), responses: { '204': { description: deleted }, '409': { description: has employees } } }
...
  /announcements:
    get: { summary: Company news, pinned first, parameters: [{ name: limit, in: query, schema: { type: integer } }], responses: { '200': { description: list } } }
    post: { summary: Publish an announcement (admin), responses: { '201': { description: created } } }
  /announcements/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    patch: { summary: Pin or unpin (admin), responses: { '200': { description: updated } } }
    delete: { summary: Delete (admin), responses: { '204': { description: deleted } } }
...